+++ packages/subscription-controller/src/types.ts
@@ -191,6 +191,7 @@ export type ISubscriptionService = {
   startSubscriptionWithCard(
     request: StartSubscriptionRequest,
   ): Promise<StartSubscriptionResponse>;
+  getBillingPortalUrl(): Promise<BillingPortalResponse>;
   getPricing(): Promise<PricingResponse>;
   startSubscriptionWithCrypto(
     request: StartCryptoSubscriptionRequest,
@@ -231,4 +232,8 @@ export type UpdatePaymentMethodCryptoRequest = {
   rawTransaction: Hex;
   recurringInterval: RecurringInterval;
   billingCycles: number;
+};
+
+export type BillingPortalResponse = {
+  url: string;
 };